# backEnd/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Documentation",
        "description": "Documentation dyal l-API b PHP 8.2 Attributes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "69a48b8c7cb177955d5c615348201c4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order",
                                    "ItemsOrder"
                                ],
                                "properties": {
                                    "order": {
                                        "required": [
                                            "user_id",
                                            "total_price",
                                            "status",
                                            "shipping_address"
                                        ],
                                        "properties": {
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "total_price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 99.99
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            },
                                            "shipping_address": {
                                                "type": "string",
                                                "example": "123 Street, Casablanca"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "ItemsOrder": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "productId",
                                                "quantity",
                                                "price"
                                            ],
                                            "properties": {
                                                "productId": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2,
                                                    "minimum": 1
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 45.5,
                                                    "minimum": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "operationId": "b9f229a2be499a1d73bc34f0d363281c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an existing order",
                "operationId": "ff1b785a826f3405f9b3ffe1691003d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an existing order",
                "operationId": "8fd202e580e5a3081f4a02e739185989",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/user/{userId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders by user ID",
                "operationId": "287ea5608b8603c2e77fc05152539fc2",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/status/{status}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders by status",
                "operationId": "044b9497990e19a075adf80367bce444",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pending"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "operationId": "f9938122690bd4abf64acec30b3eccf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "completed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Dkhel l-Token dialk hna bach t-authontifi",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}